Added RampFinder to separate concerns
